Initial website redesign.
This new version forgos js and external fonts in an effort to be a much
smaller static website. The old side navbar has been replaced with a
simpler static top navbar, and the site as a whole makes use of Nord
dark theme colors.

For now the site's content is more or less the same as the original,
except that we have a contact page instead of a direct mailto link, and
the home page image and pdf resume link have been removed.

In the future we will likely expand on the content, including a
dedicated hardware section for the projects page, as well as potentially
an entire new blog section for future writings.