Improve look of code blocks.
For now, the only noticeable difference is that the code block
backgrounds in the "Motivating Examples" section of the Unix Philosophy
blog post go all the way to the max width instead of stopping at the end
of the content.

However, we also included classes for code syntax highlighting, which
can be used for more involved programming code blocks that would read
better with a splash of color.